+++ tests/urls.py
@@ -1,9 +1,9 @@
 """ Tests App URL Config """
-from django.conf.urls import url, include
+from django.conf.urls import include
 from django.contrib import admin
-
+from django.urls import path
 
 urlpatterns = [
-    url(r'^api/password_reset/', include('django_rest_passwordreset.urls', namespace='password_reset')),
-    url(r'^admin/', admin.site.urls),
+    path("api/password_reset/", include('django_rest_passwordreset.urls', namespace='password_reset')),
+    path("admin/", admin.site.urls),
 ]