+++ tests/user_id_uuid_testapp/models.py
@@ -1,0 +1,8 @@
+import uuid
+
+from django.contrib.auth.models import AbstractUser
+from django.db import models
+
+
+class User(AbstractUser):
+    id = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)